backend/requirements.txt
alembic==1.17.2
black==25.12.0
blinker==1.9.0
cachelib==0.13.0
cffi==2.0.0
click==8.3.1
colorama==0.4.6
cryptography==46.0.3
Deprecated==1.3.1
dnspython==2.8.0
email-validator==2.3.0
Flask==3.1.2
Flask-Limiter==4.1.1
Flask-Login==0.6.3
Flask-Migrate==4.1.0
Flask-Session==0.8.0
Flask-SQLAlchemy==3.1.1
flask-talisman==1.1.0
Flask-WTF==1.2.2
greenlet==3.3.0
idna==3.11
itsdangerous==2.2.0
Jinja2==3.1.6
limits==5.6.0
Mako==1.3.10
MarkupSafe==3.0.3
msgspec==0.20.0
mypy_extensions==1.1.0
ordered-set==4.1.0
packaging==25.0
pathspec==0.12.1
platformdirs==4.5.1
pycparser==2.23
python-dotenv==1.2.1
python-http-client==3.3.7
pytokens==0.3.0
sendgrid==6.12.5
SQLAlchemy==2.0.45
typing_extensions==4.15.0
Werkzeug==3.1.4
wrapt==2.0.1
WTForms==3.2.1
